agents/career.py
"""
Career Agent - Job Search & Automated Applications
Your personal career assistant
"""

from typing import Dict, Any, List, Optional
import asyncio
import httpx
from bs4 import BeautifulSoup
from loguru import logger

class CareerAgent:
    """
    Handles job search, resume optimization, and automated applications
    """
    
    def __init__(self):
        self.job_boards = [
            "https://www.linkedin.com/jobs",
            "https://www.indeed.com",
            "https://remoteok.com/remote-jobs"
        ]
    
    async def search_jobs(self, title: str, location: str = "Remote") -> List[Dict]:
        """
        Search for jobs across multiple platforms
        """
        logger.info(f"💼 Searching jobs: {title} in {location}")
        
        jobs = []
        
        # Search RemoteOK (easy to scrape, no auth)
        try:
            url = "https://remoteok.com/remote-jobs"
            async with httpx.AsyncClient() as client:
                response = await client.get(url, timeout=10)
                soup = BeautifulSoup(response.text, 'html.parser')
                
                for job in soup.find_all('tr', class_='job')[:10]:
                    position = job.find('h2', itemprop='title')
                    company = job.find('h3', itemprop='name')
                    
                    if position and company:
                        jobs.append({
                            'title': position.text.strip(),
                            'company': company.text.strip(),
                            'location': 'Remote',
                            'url': 'https://remoteok.com' + job.get('data-url', ''),
                            'source': 'RemoteOK'
                        })
                
                logger.info(f"✅ Found {len(jobs)} jobs on RemoteOK")
                
        except Exception as e:
            logger.error(f"Job search failed: {e}")
        
        return jobs
    
    async def analyze_job(self, job: Dict, user_profile: Dict) -> Dict[str, Any]:
        """
        Analyze job fit based on user profile
        """
        # Simple keyword matching (will be enhanced with AI)
        match_score = 0.0
        
        if user_profile.get('skills'):
            job_text = f"{job['title']} {job.get('description', '')}".lower()
            matching_skills = [
                skill for skill in user_profile['skills']
                if skill.lower() in job_text
            ]
            match_score = len(matching_skills) / len(user_profile['skills'])
        
        return {
            "job": job,
            "match_score": match_score,
            "matching_skills": matching_skills if match_score > 0 else [],
            "recommendation": "Apply" if match_score > 0.5 else "Review"
        }
    
    async def auto_apply(self, job: Dict, user_profile: Dict) -> Dict[str, Any]:
        """
        Automatically apply to job (browser automation required)
        """
        logger.info(f"📝 Auto-applying to: {job['title']} at {job['company']}")
        
        # TODO: Use browser agent to fill application form
        # For now, return placeholder
        
        return {
            "status": "pending",
            "job": job['title'],
            "company": job['company'],
            "message": "Application will be processed by browser agent"
        }

# Global instance
career_agent = CareerAgent()