requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# AI & LLM
langchain==0.1.9
langchain-google-genai==0.0.11
openai==1.12.0
anthropic==0.18.1

# Browser Automation
playwright==1.42.0
selenium==4.18.1
undetected-chromedriver==3.5.5
nodriver==0.33

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
scrapy==2.11.1
requests==2.31.0
httpx==0.27.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
python-multipart==0.0.9

# Database & Storage
firebase-admin==6.4.0
redis==5.0.1
sqlalchemy==2.0.27

# Security & Authentication
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
loguru==0.7.2
python-dateutil==2.8.2
pytz==2024.1
aiofiles==23.2.1
pillow==10.2.0

# Monitoring
psutil==5.9.8